+++ FFT/AudioProcesisng/FFT.h
@@ -13,36 +13,36 @@
 // muveletek komplex szamokkal
 
 typedef struct complex_t {
-	float  re, im;
+	double  re, im;
 } complex;
 
-inline void complex_exp(float r, float phi, struct complex_t *out){
+inline void complex_exp(double r, double phi, struct complex_t *out){
 	out->re = r * cos(phi);
 	out->im = r * sin(phi);
 }
 
 inline void complex_mul(const complex *_a, const complex *_b, complex *out){
-	float a=_a->re, b=_a->im, c=_b->re, d=_b->im;
+	double a=_a->re, b=_a->im, c=_b->re, d=_b->im;
 
 	out->re = (a*c-b*d);
 	out->im = (a*d+b*c);
 }
 
-inline void complex_mul_by_c(const complex *_a, const float _b, complex *out){
-	float a=_a->re, b=_a->im;
+inline void complex_mul_by_c(const complex *_a, const double _b, complex *out){
+	double a=_a->re, b=_a->im;
 
 	out->re = _b*a;
 	out->im = _b*b;
 }
 
 inline void complex_add(const complex *_a, const complex *_b, complex *out){
-	float a=_a->re, b=_a->im, c=_b->re, d=_b->im;
+	double a=_a->re, b=_a->im, c=_b->re, d=_b->im;
 	out->re = a+c;
 	out->im = b+d;
 }
 
 inline void complex_sub(const  complex *_a, const complex *_b, complex *out){
-	float a=_a->re, b=_a->im, c=_b->re, d=_b->im;
+	double a=_a->re, b=_a->im, c=_b->re, d=_b->im;
 	out->re = a-c;
 	out->im = b-d;
 }
@@ -61,7 +61,7 @@ inline void icomplex_mul(const intComplex *_a, const intComplex *_b, intComplex 
 	out->im = (a*d+b*c);
 }
 
-inline void icomplex_mul_by_c(const intComplex *_a, const float _b, intComplex *out){
+inline void icomplex_mul_by_c(const intComplex *_a, const double _b, intComplex *out){
 	int a=_a->re, b=_a->im;
 
 	out->re = _b*a;
@@ -81,7 +81,7 @@ inline void icomplex_sub(const  intComplex *_a, const intComplex *_b, intComplex
 }
 
 //////////////////////////////////////////////////////////////////////////////
-// 32-bit Float Fast FT
+// 32-bit double Fast FT
 /**
 */
 class FFT {
@@ -91,13 +91,13 @@ class FFT {
 
 		// CFFTv2 1:1-ben bekerul ide.
 		inline void fft_ccopy ( int n, complex *x, complex *y ){  int i; for (i=0; i<n; i++ ) y[i] = x[i]; }
-		void step (int n, int mj, complex *a, complex *b, complex *c, complex *d, float sgn);
+		void step (int n, int mj, complex *a, complex *b, complex *c, complex *d, double sgn);
 
 	public:
 		FFT(int samplesize, int samplerate);
 		virtual ~FFT();
 
-		void calculate(float *inbuf);
+		void calculate(double *inbuf);
 		void calculate();
 
 		inline complex *getInputBuffer(){return this->x;}
